Revert "Updated SQLite3 serial column to not use AUTOINCREMENT"
This reverts commit aaea8984b827bd205c26d774d40a342c017bc498.

I was mistaken with this commit.  AUTOINCREMENT should be used and not
the default behavior of SQLite3, which is to increment the integer
primary key.  The problem with that approach is that once all the keys
are used up, SQLite3 will begin to reuse keys for records that have been
deleted, which is definately a bad idea in most cases.